nan fixed

+++ src/pkts.h
@@ -6,7 +6,7 @@ struct rcv_pkt {
   // jigdump header 
   int8_t rssi;
   u_int8_t channel_rcv;
-  float rate;
+  u_int16_t rate;
   u_int16_t  freq ;
   u_int32_t ath_crc_err;
   u_int32_t ath_phy_err;
